fix(frontend): fix storico protocolli  e visualliza lista protocolli cliente
