+++ frontend/src/App.jsx
@@ -17,7 +17,7 @@ import {AuthProvider} from "./context/AuthContext";
 import AppShell from "./AppShell";
 import Dashboard from "./pages/Dashboard";
 import {FetchProvider} from "./context/FetchContext";
-import TProtocolsList from "./pages/User/Trainer/TProtocolsList";
+
 
 const AppRoutes = () => {
     return (
@@ -32,15 +32,14 @@ const AppRoutes = () => {
             <Route path="account" element={<AppShell><Edit/></AppShell>}/>
             <Route path="/customer" element={<AppShell><CustomerIndex/></AppShell>}>
                 <Route path="create" element={<Create/>}/>
-                <Route path="protocol/:id" element={<CustomerviewProtocol/>}/>
-                <Route path="protocols" element={ <ProtocolsList/>}/>
             </Route>
-
+            <Route path="protocols" element={<AppShell><ProtocolsList/></AppShell>}/>
+            <Route path="protocols/:id" element={<AppShell><CustomerviewProtocol/></AppShell>}/>
             <Route path="/trainer" element={<AppShell><TrainerIndex/></AppShell>}>
                 <Route path="edit" element={<AppShell><Edit/></AppShell>}/>
                 <Route path="addCustomer" element={<Create/>}/>
-                <Route path="protocols/:id" element={<TProtocolsList/>}/>
             </Route>
+
         </Routes>
     )
 }
